% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseShortName.R
\name{chooseShortName}
\alias{chooseShortName}
\title{Short algorithm to choose a short name based on a given long name}
\usage{
chooseShortName(
  fullname = stop("Invalid string name entered."),
  withPrefix = NULL,
  withSuffix = NULL,
  envir = NULL,
  silent = F
)
}
\arguments{
\item{fullname}{The name you intend to shorten}

\item{withPrefix}{Prefix to include in front of the new short name}

\item{withSuffix}{Suffix to include in front of the new short name}

\item{envir}{The environment where to store the name}

\item{silent}{Return response at the end of evaluations?}
}
\description{
A very simple and short algorithm to choose a short name based on a given name
}
