% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseShortName.R
\name{clearStoredNames}
\alias{clearStoredNames}
\title{Clears previously stored names}
\usage{
clearStoredNames()
}
\description{
Beware that by clearing all stored names, you may inadventently duplicate new names
}
