% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasSpecialCharacters.R
\name{hasSpecialCharacters}
\alias{hasSpecialCharacters}
\title{Does string have special characters?}
\usage{
hasSpecialCharacters(string)
}
\arguments{
\item{string}{The string to evaluate}
}
\description{
Evaluates if a specified string contains special characters
}
