% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasUpperCase.R
\name{hasUpperCase}
\alias{hasUpperCase}
\title{Contains uppercase?}
\usage{
hasUpperCase(string)
}
\arguments{
\item{string}{the string to evaluate}
}
\description{
Simply change if there is any uppercase letter in a string
}
