% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{help}
\alias{help}
\title{Help for all functions}
\usage{
help(functionName, package = NULL)
}
\arguments{
\item{functionName}{function name to search for}

\item{package}{package name containing the function. Leave unentered if this is a shorthand function}
}
\value{
Documentation for function
}
\description{
Access help for all functions including any newly created shorthand functions
}
\examples{
\donttest{
pkgname <- 'qpdf' #pacakge name
shortenPkg(pkgname,TRUE) #shorten the package
index(pkgname) #index the package functions shortened
help('qpd.pl') #choose a function name and find help
}

}
