% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexPkg.R
\name{index}
\alias{index}
\title{Index a shortened package}
\usage{
index(pkg = stop("Enter a package name to index"))
}
\arguments{
\item{pkg}{The package name}
}
\description{
This function will provide you the list of shorthand functions created for a package
}
