% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isUpperCase.R
\name{isUpperCase}
\alias{isUpperCase}
\title{Is string uppercase?}
\usage{
isUpperCase(string)
}
\arguments{
\item{string}{The string to evaluate}
}
\description{
Simply, evaluate if a string is uppercase
}
