% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2shortcode-package.R
\docType{package}
\name{r2shortcode-package}
\alias{r2shortcode}
\alias{r2shortcode-package}
\title{r2shortcode: Shorten Function Names of Functions in Another Package and Create an Index to Make Them Accessible}
\description{
When creating a package, authors may sometimes struggle with coming up with easy and straightforward function names, and at the same time hoping that other packages do not already have the same function names. In trying to meet this goal, sometimes, function names are not descriptive enough and may confuse the potential users. The purpose of this package is to serve as a package function short form generator and also provide shorthand names for other functions. Having this package will entice authors to create long function names without the fear of users not wanting to use their packages because of the long names. In a way, everyone wins - the authors can use long descriptive function names, and the users can use this package to make short functions names while still using the package in question.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oobianom/r2shortcode}
  \item Report bugs at \url{https://github.com/oobianom/r2shortcode}
}

}
\keyword{internal}
