% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchNameSaveName.R
\name{searchNameSaveName}
\alias{searchNameSaveName}
\title{Search for the existence of a name and save if unavailable}
\usage{
searchNameSaveName(withPrefix = NULL, strN.complete, withSuffix = NULL)
}
\arguments{
\item{withPrefix}{Provide the prefix to use for the function name, could use NULL}

\item{strN.complete}{Provide the function name to search for, and save if feasible}

\item{withSuffix}{Provide the suffix to use for the function name, could use NULL}
}
\value{
Concatenate of search and a boolean
}
\description{
Carries out a search on the already used shorthand function names and if the new name does not exist, it saves it
}
\examples{
nameToCheck <- 'ObiObianom'
searchNameSaveName(strN.complete= nameToCheck)
searchNameSaveName(strN.complete= nameToCheck)


}
