% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{whatis}
\alias{whatis}
\title{Help for all functions}
\usage{
whatis(functionName, package = NULL)
}
\arguments{
\item{functionName}{function name to search for}

\item{package}{package name containing the function. Leave unentered if this is a shorthand function}
}
\description{
Access help for all functions including any newly created shorthand functions
}
