% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discardShortcodes.R
\name{discardShortcodes}
\alias{discardShortcodes}
\title{Discard shortcodes}
\usage{
discardShortcodes(
  pkg,
  reflib = options()$.funCNamesPkgReference,
  response = TRUE
)
}
\arguments{
\item{pkg}{package name}

\item{reflib}{reference library, preferrably leave unentered}

\item{response}{TRUE or FALSE, return a response upon completion}
}
\value{
unloads short function names
}
\description{
This will discard all shorthand functions created and delete them from stores as well. Good if you inadvertently shorthand a package.
}
\examples{
\donttest{
pkgName = 'quickcode'
shortenPkg(pkgName)
discardShortcodes(pkgName)
}

}
