% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasSpecialCharacters.R
\name{hasSpecialCharacters}
\alias{hasSpecialCharacters}
\title{Does string have special characters?}
\usage{
hasSpecialCharacters(string)
}
\arguments{
\item{string}{The string to evaluate}
}
\value{
TRUE or FALSE
}
\description{
Evaluates if a specified string contains special characters
}
\examples{
strToTest1 <- 'obi_\%_good_^you'
strToTest2 <- 'obigoodyou'
hasSpecialCharacters(strToTest1)
hasSpecialCharacters(strToTest2)

}
