% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isUpperCase.R
\name{isUpperCase}
\alias{isUpperCase}
\title{Is string uppercase?}
\usage{
isUpperCase(string)
}
\arguments{
\item{string}{The string to evaluate}
}
\value{
TRUE or FALSE if the string is all upper case
}
\description{
Simply, test if a string is uppercase and return TRUE or FALSE
}
\examples{
strTest1 <- 'OBI'
strTest2 <- 'obiO'
isUpperCase(strTest1)
isUpperCase(strTest2)


}
