% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storeChosenName.R
\name{storeChosenName}
\alias{storeChosenName}
\title{Store chosen name in memory}
\usage{
storeChosenName(name)
}
\arguments{
\item{name}{The chosen name to store}
}
\value{
store the chosen name in storage variable
}
\description{
When all else passes, store this particular name in memory
}
\examples{
nameStore <- 'ObiStore1'
storeChosenName(nameStore)

}
